requirements.txt
streamlit>=1.26.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
protobuf>=4.23.0
